fix(plugins): remove protobuf duplications (#2729)
* fix(plugins): remove protobuf duplications

* style(fmt): rustfmt

+++ zellij-tile/src/shim.rs
@@ -133,7 +133,7 @@ pub fn switch_tab_to(tab_idx: u32) {
 
 /// Set a timeout in seconds (or fractions thereof) after which the plugins [update](./plugin-api-events#update) method will be called with the [`Timer`](./plugin-api-events.md#timer) event.
 pub fn set_timeout(secs: f64) {
-    let plugin_command = PluginCommand::SetTimeout(secs as f32);
+    let plugin_command = PluginCommand::SetTimeout(secs);
     let protobuf_plugin_command: ProtobufPluginCommand = plugin_command.try_into().unwrap();
     object_to_stdout(&protobuf_plugin_command.encode_to_vec());
     unsafe { host_run_plugin_command() };
